Added Replying Kafka comms, Logging and UUIDs

+++ WitChallenge/common/src/main/java/com/example/requests/CalculatorRequest.java
@@ -1,15 +1,17 @@
 package com.example.requests;
 
+import java.math.BigDecimal;
+
 public class CalculatorRequest {
 
     private String operation;
-    private int value1;
-    private int value2;
+    private BigDecimal value1;
+    private BigDecimal value2;
 
     public CalculatorRequest() {
     }
 
-    public CalculatorRequest(String operation, int value1, int value2) {
+    public CalculatorRequest(String operation, BigDecimal value1, BigDecimal value2) {
         this.operation = operation;
         this.value1 = value1;
         this.value2 = value2;
@@ -19,11 +21,11 @@ public class CalculatorRequest {
         return operation;
     }
 
-    public int getValue1() {
+    public BigDecimal getValue1() {
         return value1;
     }
 
-    public int getValue2() {
+    public BigDecimal getValue2() {
         return value2;
     }
 
@@ -31,11 +33,11 @@ public class CalculatorRequest {
         this.operation = operation;
     }
 
-    public void setValue1(int value1) {
+    public void setValue1(BigDecimal value1) {
         this.value1 = value1;
     }
 
-    public void setValue2(int value2) {
+    public void setValue2(BigDecimal value2) {
         this.value2 = value2;
     }
 }
